requirements.txt
Flask >= 3.0.3
tensorflow >= 2.16.0
pillow >= 10.4.0
ngrok >= 1.4.0